Move Visitor as a top level concern of Spoom
